[TASK] Refactor `HistoryRecord` to `AbstractEntityEventListener`
The original idea was to make it possible to use the trait in other
Classes beside to the eventListeners. As this makes no sense at the
moment, move the functionality to an abstract class and the single
functionality to thhe respective EventListeners.

As an extension of EventListeners is not useful, make them final.

Add DocBlocks to the classes describing their use case.

Remove the `RecordHistoryStore`, which was a copy of the original one
from the core, as this could handle our issues, too.